+++ src/blocks/Form/__stories__/Form.stories.tsx
@@ -5,6 +5,7 @@ import {blockTransform} from '../../../../.storybook/utils';
 import {FormBlockDirection, FormBlockModel, FormBlockProps} from '../../../models';
 import FormBlock from '../Form';
 
+import ExampleAdditionContent, {ContentLabelProps} from './components/ExampleAdditionalContent';
 import ExampleStub from './components/ExmapleStub';
 
 import data from './data.json';
@@ -45,6 +46,7 @@ export const WithBackgroundImage = VariantsTemplate.bind([]);
 export const DarkTheme = VariantsTemplate.bind([]);
 export const FormData = VariantsTemplate.bind([]);
 export const WithCustomFormNode = DefaultTemplate.bind([]);
+export const WithAdditionalContentNode = VariantsTemplate.bind([]);
 
 Default.args = data.default as FormBlockModel;
 
@@ -93,3 +95,10 @@ FormData.parameters = {
 };
 
 WithCustomFormNode.args = {...data.default, customFormNode: <ExampleStub />} as FormBlockModel;
+
+WithAdditionalContentNode.args = DIRECTIONS_VARIANTS.map((variant) => ({
+    ...variant,
+    additionalContentNode: (
+        <ExampleAdditionContent labels={data.withAdditionalContent.labels as ContentLabelProps[]} />
+    ),
+})) as FormBlockModel[];