+++ src/components/Buttons/Buttons.tsx
@@ -1,3 +1,5 @@
+import * as React from 'react';
+
 import {ButtonProps, ContentSize} from '../../models';
 import {block} from '../../utils';
 import Button from '../Button/Button';
@@ -8,7 +10,7 @@ const b = block('buttons');
 
 type ButtonsProps = {
     className?: string;
-    buttons?: ButtonProps[];
+    buttons?: (ButtonProps | React.ReactNode)[];
     size?: ContentSize;
     titleId?: string;
     qa?: string;
@@ -19,7 +21,9 @@ function getButtonSize(size: ContentSize) {
     switch (size) {
         case 's':
             return 'm';
+        case 'm':
         case 'l':
+        case 'xl':
         default:
             return 'xl';
     }
@@ -28,19 +32,25 @@ function getButtonSize(size: ContentSize) {
 const Buttons = ({className, titleId, buttons, size = 's', qa, buttonQa}: ButtonsProps) =>
     buttons ? (
         <div className={b({size}, className)} data-qa={qa}>
-            {buttons.map((item) => (
-                <Button
-                    className={b('button')}
-                    {...item}
-                    key={item.url}
-                    size={getButtonSize(size)}
-                    qa={buttonQa}
-                    extraProps={{
-                        'aria-describedby': item.urlTitle ? undefined : titleId,
-                        ...item.extraProps,
-                    }}
-                />
-            ))}
+            {buttons.map((item, index) =>
+                React.isValidElement(item) ? (
+                    <React.Fragment key={index}>{item}</React.Fragment>
+                ) : (
+                    <Button
+                        className={b('button')}
+                        {...(item as ButtonProps)}
+                        key={(item as ButtonProps).url}
+                        size={getButtonSize(size)}
+                        qa={buttonQa}
+                        extraProps={{
+                            'aria-describedby': (item as ButtonProps).urlTitle
+                                ? undefined
+                                : titleId,
+                            ...(item as ButtonProps).extraProps,
+                        }}
+                    />
+                ),
+            )}
         </div>
     ) : null;