+++ src/components/Links/Links.tsx
@@ -12,7 +12,9 @@ function getLinkSize(size: ContentSize) {
     switch (size) {
         case 's':
             return 'm';
+        case 'm':
         case 'l':
+        case 'xl':
         default:
             return 'l';
     }