+++ src/gravity-blocks/schema/validators/common.ts
@@ -14,7 +14,7 @@ export const mediaDirection = ['media-content', 'content-media'];
 export const textSize = ['xs', 's', 'sm', 'm', 'l'];
 export const containerSizesArray = ['sm', 'md', 'lg', 'xl', 'all'];
 export const sliderSizesArray = ['sm', 'md', 'lg', 'xl'];
-export const contentSizes = ['s', 'm', 'l'];
+export const contentSizes = ['s', 'm', 'l', 'xl'];
 export const contentTextWidth = ['s', 'm', 'l'];
 
 export const videoTypes = ['default', 'player'];
@@ -800,4 +800,30 @@ export const GravityIconProps = {
             },
         },
     ],
+};
+
+export const HeaderBreadcrumbsProps = {
+    type: 'object',
+    additionalProperties: false,
+    required: ['items'],
+    properties: {
+        items: {
+            type: 'array',
+            items: {
+                type: 'object',
+                additionalProperties: false,
+                required: ['url', 'text'],
+                properties: {
+                    url: {
+                        type: 'string',
+                    },
+                    text: {
+                        type: 'string',
+                        contentType: 'text',
+                    },
+                },
+            },
+        },
+        theme: {type: 'string', enum: ['light', 'dark']},
+    },
 };