+++ src/models/constructor-items/blocks.ts
@@ -5,7 +5,7 @@ import {ButtonSize} from '@gravity-ui/uikit';
 import {GridColumnSize, GridColumnSizesType, IndentValue} from '../../gravity-blocks/grid/types';
 import {ThemeSupporting} from '../../utils';
 import {DeviceSupporting} from '../../utils/breakpoint';
-import {AnalyticsEventsBase} from '../common';
+import {AnalyticsEventsBase, AnalyticsEventsProp} from '../common';
 
 import {
     AnchorProps,
@@ -28,6 +28,8 @@ import {
     LegendTableMarkerType,
     LinkProps,
     MapProps,
+    MediaComponentImageProps,
+    MediaComponentVideoProps,
     MediaDirection,
     MediaProps,
     MediaView,
@@ -59,6 +61,7 @@ export enum BlockType {
     TabsBlock = 'tabs-block',
     HeaderSliderBlock = 'header-slider-block',
     HeaderBlock = 'header-block',
+    HeroBlock = 'hero-block',
     IconsBlock = 'icons-block',
     CardLayoutBlock = 'card-layout-block',
     ContentLayoutBlock = 'content-layout-block',
@@ -69,7 +72,11 @@ export enum BlockType {
 }
 
 export const BlockTypes = Object.values(BlockType);
-export const HeaderBlockTypes = [BlockType.HeaderBlock, BlockType.HeaderSliderBlock];
+export const HeaderBlockTypes = [
+    BlockType.HeaderBlock,
+    BlockType.HeaderSliderBlock,
+    BlockType.HeroBlock,
+];
 
 export interface Childable {
     children?: SubBlock[];
@@ -201,11 +208,35 @@ export interface HeaderBlockProps {
     renderTitle?: (title: string) => React.ReactNode;
 }
 
-export interface ExtendedFeaturesItem
-    extends Omit<
+export interface HeroBlockBackground
+    extends Partial<MediaComponentImageProps>,
+        Partial<MediaComponentVideoProps> {
+    color?: string;
+}
+
+export interface HeroBlockMedia extends Partial<MediaProps> {
+    roundCorners?: boolean;
+}
+
+export interface HeroBlockProps
+    extends Pick<
         ContentBlockProps,
-        'theme' | 'centered' | 'colSizes' | 'size' | 'title' | 'labels'
+        'title' | 'text' | 'list' | 'additionalInfo' | 'links' | 'theme'
     > {
+    breadcrumbs?: HeaderBreadCrumbsProps;
+    // TODO: add overtitle to ContentProps
+    overtitle?: string | JSX.Element;
+    buttons?: ThemeSupporting<
+        Pick<ButtonProps, 'url' | 'text' | 'theme' | 'primary' | 'extraProps'> | React.ReactNode
+    >[];
+    media?: ThemeSupporting<HeroBlockMedia>;
+    fullWidth?: boolean;
+    verticalOffset?: 's' | 'm' | 'l' | 'xl';
+    background?: ThemeSupporting<HeroBlockBackground>;
+}
+
+export interface ExtendedFeaturesItem
+    extends Omit<ContentBlockProps, 'theme' | 'centered' | 'colSizes' | 'size' | 'title'> {
     title: string;
     label?: string;
     icon?: ThemedImage;
@@ -243,7 +274,7 @@ export interface QuestionItem {
 }
 
 export interface QuestionsProps
-    extends Omit<ContentBlockProps, 'colSizes' | 'centered' | 'size' | 'theme' | 'labels'> {
+    extends Omit<ContentBlockProps, 'colSizes' | 'centered' | 'size' | 'theme'> {
     items: QuestionItem[];
 }
 
@@ -260,7 +291,7 @@ export interface FoldableListItem {
 }
 
 export interface FoldableListProps
-    extends Omit<ContentBlockProps, 'colSizes' | 'centered' | 'size' | 'theme' | 'labels'> {
+    extends Omit<ContentBlockProps, 'colSizes' | 'centered' | 'size' | 'theme'> {
     items: FoldableListItem[];
 }
 
@@ -286,7 +317,7 @@ export interface MediaBaseBlockProps extends Animatable, MediaContentProps {
 }
 
 export interface MediaContentProps
-    extends Omit<ContentBlockProps, 'colSizes' | 'text' | 'theme' | 'centered' | 'labels'> {
+    extends Omit<ContentBlockProps, 'colSizes' | 'text' | 'theme' | 'centered'> {
     description?: string;
     /** @deprecated  Use array of buttons from ContentBlockProps instead**/
     button?: ButtonProps;
@@ -311,8 +342,8 @@ export interface InfoBlockProps {
     sectionsTitle?: string;
     /** @deprecated **/
     links?: Pick<LinkProps, 'text' | 'url'>[];
-    leftContent?: Omit<ContentBlockProps, 'colSizes' | 'theme' | 'size' | 'labels'>;
-    rightContent?: Omit<ContentBlockProps, 'colSizes' | 'theme' | 'size' | 'labels'>;
+    leftContent?: Omit<ContentBlockProps, 'colSizes' | 'theme' | 'size'>;
+    rightContent?: Omit<ContentBlockProps, 'colSizes' | 'theme' | 'size'>;
 }
 
 export interface TableProps {
@@ -333,7 +364,7 @@ export interface TableBlockProps {
 }
 
 export interface TabsBlockItem
-    extends Omit<ContentBlockProps, 'size' | 'colSizes' | 'centered' | 'theme' | 'labels'>,
+    extends Omit<ContentBlockProps, 'size' | 'colSizes' | 'centered' | 'theme'>,
         WithBorder {
     tabName: string;
     /**
@@ -371,6 +402,7 @@ export interface CardLayoutBlockProps extends Childable, Animatable, LoadableChi
 export type FilterTag = {
     id: string;
     label: string;
+    analyticsEvent?: AnalyticsEventsProp;
 };
 
 export type FilterItem = {
@@ -384,7 +416,7 @@ export interface FilterBlockProps extends Animatable {
     title?: TitleItemProps | string;
     description?: string;
     tagButtonSize?: ButtonSize;
-    allTag?: boolean | string;
+    allTag?: boolean | string | FilterTag;
     colSizes?: GridColumnSizesType;
     centered?: boolean;
 }
@@ -412,7 +444,7 @@ interface ContentLayoutBlockParams {
 }
 
 export interface ContentLayoutBlockProps extends ContentLayoutBlockParams {
-    textContent: Omit<ContentBlockProps, 'labels'>;
+    textContent: ContentBlockProps;
     fileContent?: FileLinkProps[];
 }
 
@@ -431,18 +463,6 @@ export interface ContentListProps {
     theme?: ContentTheme;
 }
 
-export interface ContentLabelProps {
-    text: string;
-    icon?: ThemeSupporting<ImageProps | SVGIcon>;
-    gravityIcon?: ThemeSupporting<GravityIconProps>;
-}
-
-export interface ContentLabelsProps {
-    labels: ContentLabelProps[];
-    size?: ContentSize;
-    theme?: ContentTheme;
-}
-
 export interface ContentBlockProps {
     title?: TitleItemBaseProps | string;
     titleId?: string;
@@ -450,13 +470,12 @@ export interface ContentBlockProps {
     textId?: string;
     additionalInfo?: string;
     links?: LinkProps[];
-    buttons?: ButtonProps[];
+    buttons?: (ButtonProps | React.ReactNode)[];
     size?: ContentSize;
     colSizes?: GridColumnSizesType;
     centered?: boolean;
     theme?: ContentTheme;
     list?: ContentItemProps[];
-    labels?: ContentLabelProps[];
     controlPosition?: 'default' | 'bottom';
 }
 
@@ -505,12 +524,17 @@ export interface FormBlockProps {
     direction?: FormBlockDirection;
     background?: ThemeSupporting<FormBlockBackgroundProps>;
     customFormNode?: React.ReactNode;
+    additionalContentNode?: React.ReactNode;
 }
 
 //block models
 export type HeaderBlockModel = {
     type: BlockType.HeaderBlock;
 } & HeaderBlockProps;
+
+export type HeroBlockModel = {
+    type: BlockType.HeroBlock;
+} & HeroBlockProps;
 
 export type SliderOldBlockModel = {
     type: BlockType.SliderOldBlock;
@@ -607,6 +631,7 @@ type BlockModels =
     | TableBlockModel
     | TabsBlockModel
     | HeaderBlockModel
+    | HeroBlockModel
     | IconsBlockModel
     | HeaderSliderBlockModel
     | CardLayoutBlockModel